reformat the display services implementation
