+++ src/main/kotlin/org/jetbrains/research/testspark/actions/llm/LLMSampleSelectorBuilder.kt
@@ -226,9 +226,9 @@ class LLMSampleSelectorBuilder(private val project: Project, private val languag
         var normalizedImports = imports
         if (normalizedImports.isNotBlank()) normalizedImports += "\n\n"
         return normalizedImports +
-                "public class TestSample {\n" +
-                "   $methodCode\n" +
-                "}"
+            "public class TestSample {\n" +
+            "   $methodCode\n" +
+            "}"
     }
 
     private fun createMethodName(psiClass: PsiClass, method: PsiMethod): String =
