+++ src/main/kotlin/org/jetbrains/research/testspark/helpers/CoverageHelper.kt
@@ -16,7 +16,7 @@ import com.intellij.ui.components.JBLabel
 import com.intellij.ui.components.JBScrollPane
 import com.intellij.util.ui.FormBuilder
 import org.jetbrains.research.testspark.services.EvoSuiteSettingsService
-import org.jetbrains.research.testspark.services.java.JavaTestCaseDisplayBuilder
+import org.jetbrains.research.testspark.services.TestCaseDisplayBuilder
 import org.jetbrains.research.testspark.settings.evosuite.EvoSuiteSettingsState
 import java.awt.Color
 import java.awt.Dimension
@@ -130,7 +130,7 @@ class CoverageHelper(
      * @param name name of the test to highlight
      */
     private fun highlightInToolwindow(name: String) {
-        val testCaseDisplayService = project.service<JavaTestCaseDisplayBuilder>()
+        val testCaseDisplayService = project.service<TestCaseDisplayBuilder>()
 
         testCaseDisplayService.highlightTestCase(name)
     }
@@ -141,7 +141,7 @@ class CoverageHelper(
      * @param map map of mutant operations -> List of names of tests which cover the mutants
      */
     private fun highlightMutantsInToolwindow(mutantName: String, map: HashMap<String, MutableList<String>>) {
-        val testCaseDisplayService = project.service<JavaTestCaseDisplayBuilder>()
+        val testCaseDisplayService = project.service<TestCaseDisplayBuilder>()
 
         testCaseDisplayService.highlightCoveredMutants(map.getOrPut(mutantName) { ArrayList() })
     }
