+++ src/main/kotlin/org/jetbrains/research/testspark/tools/ToolUtils.kt
@@ -15,8 +15,8 @@ import org.jetbrains.research.testspark.core.test.SupportedLanguage
 import org.jetbrains.research.testspark.core.utils.DataFilesUtil
 import org.jetbrains.research.testspark.data.IJTestCase
 import org.jetbrains.research.testspark.services.TestsExecutionResultService
-import java.io.File
 import org.jetbrains.research.testspark.testmanager.TestGeneratorFactory
+import java.io.File
 
 object ToolUtils {
     val sep = File.separatorChar
